port(EffectsCore): bring fxcurve.cpp into builds
24 upstream files total in the build. fxcurve.cpp is the FxCurve sampler
used by the particle effects subsystem.

Source fixes:
- src/EffectsCore/fxcurve.cpp: 4 sign-compare sites between
  source->currentKeyIndex (int32_t) and (uint32_t)keyCount. Cast LHS to
  uint32_t to match — keyIndex is always non-negative in upstream usage.

Stub:
- Hunk_AllocAlign in posix_stubs.cpp: same backing as Hunk_Alloc (new[]
  satisfies alignof(std::max_align_t) which covers EffectsCore's float
  array needs). Replaced with proper aligned bump when the real hunk
  subsystem ports.

The rest of EffectsCore (fx_dvars, fx_graph, fx_update_util) compiles
clean but pulls dvar.cpp's RegisterInt/Bool/Float plus math helpers
(FX_RandomDir, UnitQuatToAxis, Vec4IsNormalized, Vec3Basis_RightHanded)
and the fx_randomTable global. They land alongside dvar.

+++ src/EffectsCore/fxcurve.cpp
@@ -179,7 +179,7 @@ const FxCurve *__cdecl FxCurve_AllocAndCreateWithKeys(float *keyArray, int32_t d
 void __cdecl FxCurveIterator_SampleTimeVec3(FxCurveIterator *source, float *replyVector, float time)
 {
     FxCurveIterator_MoveToTime(source, time);
-    if (source->currentKeyIndex >= (uint32_t)(source->master->keyCount - 1))
+    if ((uint32_t)source->currentKeyIndex >= (uint32_t)(source->master->keyCount - 1))
         MyAssertHandler(
             "c:\\trees\\cod3\\src\\effectscore\\FxCurve.h",
             148,
@@ -193,7 +193,7 @@ void __cdecl FxCurveIterator_SampleTimeVec3(FxCurveIterator *source, float *repl
 double __cdecl FxCurveIterator_SampleTime(FxCurveIterator *source, float time)
 {
     FxCurveIterator_MoveToTime(source, time);
-    if (source->currentKeyIndex >= (uint32_t)(source->master->keyCount - 1))
+    if ((uint32_t)source->currentKeyIndex >= (uint32_t)(source->master->keyCount - 1))
         MyAssertHandler(
             "c:\\trees\\cod3\\src\\effectscore\\FxCurve.h",
             126,
@@ -240,7 +240,7 @@ void __cdecl FxCurveIterator_MoveToTime(FxCurveIterator *source, float time)
             "%s\n\t(source->master->dimensionCount) = %i",
             "(source->master->dimensionCount > 0)",
             source->master->dimensionCount);
-    if (source->currentKeyIndex >= (uint32_t)source->master->keyCount)
+    if ((uint32_t)source->currentKeyIndex >= (uint32_t)source->master->keyCount)
         MyAssertHandler(
             "c:\\trees\\cod3\\src\\effectscore\\FxCurve.h",
             91,
@@ -271,7 +271,7 @@ void __cdecl FxCurveIterator_MoveToTime(FxCurveIterator *source, float time)
             "key == &source->master->keys[source->currentKeyIndex * keySize]",
             v2);
     }
-    if (source->currentKeyIndex >= (uint32_t)source->master->keyCount)
+    if ((uint32_t)source->currentKeyIndex >= (uint32_t)source->master->keyCount)
         MyAssertHandler(
             "c:\\trees\\cod3\\src\\effectscore\\FxCurve.h",
             109,
